backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Werkzeug==3.1.6
pandas==2.3.3
requests==2.32.5
google-generativeai==0.8.6
python-dotenv==1.2.2
openpyxl==3.1.2
xlrd==2.0.2
python-docx==1.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
